Rename: group api 구조 수정
